refactor(files): rename files and remove unused
