--- src/main/factories/usecases/convert-number-to-full-text-usecase-factory.ts
@@ -1,5 +1,0 @@
-import { ConvertNumberToFullText } from '../../../application/ConvertNumberToFullText';
-
-export const makeConvertNumberToFullText = (): ConvertNumberToFullText => {
-  return new ConvertNumberToFullText();
-};
