--- src/main/factories/validators/make-number-validator-factory.ts
@@ -1,6 +1,0 @@
-import { Validator } from '../../../presentation/protocols';
-import { NumberInputValidator } from '../../../presentation/validation/NumberInputValidator';
-
-export const makeNumberValidator = (): Validator => {
-  return new NumberInputValidator();
-};
