--- src/presentation/protocols/controller.ts
@@ -1,5 +1,0 @@
-import { HttpRequest, HttpResponse } from '../http';
-
-export interface Controller {
-  handle(httpRequest: HttpRequest): Promise<HttpResponse>;
-}
