+++ src/presentation/protocols/index.ts
@@ -1,2 +1,1 @@
-export * from './controller';
-export * from './validator';
+export * from './Validator';
